only delete the original UserSignup if its status is set to Deactivated

+++ controllers/usersignup/usersignup_controller.go
@@ -136,7 +136,15 @@ func (r *Reconciler) Reconcile(ctx context.Context, request ctrl.Request) (ctrl.
 			Name:      originalUserSignupName,
 		}, userSignupToDelete)
 		if err == nil {
-			// Delete the original UserSignup
+			// Delete the original UserSignup, if it has finished deactivating
+			cond, found := condition.FindConditionByType(userSignupToDelete.Status.Conditions, toolchainv1alpha1.UserSignupComplete)
+			if !found || cond.Reason != toolchainv1alpha1.UserSignupUserDeactivatedReason {
+				// The UserSignup to delete isn't finished deactivating yet, return an error
+				return reconcile.Result{}, r.wrapErrorWithStatusUpdate(logger, userSignup, r.setStatusMigrationFailedCleanup,
+					errs.New("Original UserSignup not deactivated"),
+					fmt.Sprintf("Original UserSignup [%s] not yet deactivated", userSignupToDelete.Name))
+			}
+
 			err = r.Client.Delete(context.TODO(), userSignupToDelete)
 			if err != nil {
 				return reconcile.Result{}, r.wrapErrorWithStatusUpdate(logger, userSignup, r.setStatusMigrationFailedCleanup,
